requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0

# Deep Learning & RL
torch==2.0.1
stable-baselines3==2.1.0
gymnasium==0.29.0

# Data
yfinance==0.2.28

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Configuration
pyyaml==6.0.1

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.2

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2

# Optional (for streamlit app)
streamlit==1.26.0
plotly==5.16.1

tensorboard
